Allow Puma to bind to unix socket also in production
Previous to this commit the container image had the
`PORTUS_PUMA_HOST` already set.

That preventend deployments with Puma binding to a
unix socket because `config/puma.rb` resorts to them
only when `PORTUS_PUMA_HOST` is not set.

This doesn't break the compose deployment because this one
already specifies a value for `PORTUS_PUMA_HOST`.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>